# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management API",
        "description": "API for managing school operations including students, teachers, courses, grades, and attendance",
        "contact": {
            "email": "admin@school.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/docentes": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Obtiene la lista de docentes con opciones de filtro.",
                "operationId": "58c30b07fcdcd5cfe6dcb704e523a33c",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Filtrar por Nombre completo del docente.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de docentes obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre_completo": {
                                                "type": "string",
                                                "example": "Carlos Garcia"
                                            },
                                            "telefono": {
                                                "type": "string",
                                                "example": "111-222-3333"
                                            },
                                            "correo_electronico": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "carlos.garcia@example.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Crea un nuevo docente.",
                "operationId": "704947e2a274062348a0f56aaab046e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "identification_number",
                                    "gender",
                                    "date_of_birth",
                                    "phone",
                                    "address",
                                    "department",
                                    "hire_date",
                                    "emergency_contact_name",
                                    "emergency_contact_phone",
                                    "emergency_contact_email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Garcia"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos.garcia@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "identification_number": {
                                        "type": "string",
                                        "example": "987-654321-0"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1980-03-15"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "111-222-3333"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Avenida Siempre Viva 742"
                                    },
                                    "profile_image": {
                                        "description": "Optional: Teacher's profile image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "department": {
                                        "type": "string",
                                        "example": "Matemáticas"
                                    },
                                    "hire_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2010-08-01"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "50000.00"
                                    },
                                    "qualification": {
                                        "type": "string",
                                        "example": "PhD en Matemáticas"
                                    },
                                    "specialization": {
                                        "type": "string",
                                        "example": "Álgebra Lineal"
                                    },
                                    "office_location": {
                                        "type": "string",
                                        "example": "Edificio A, Oficina 201"
                                    },
                                    "emergency_contact_name": {
                                        "type": "string",
                                        "example": "Ana Garcia"
                                    },
                                    "emergency_contact_phone": {
                                        "type": "string",
                                        "example": "555-987-6543"
                                    },
                                    "emergency_contact_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ana.garcia@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Docente creado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Teacher created successfully"
                                        },
                                        "teacher": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre_completo": {
                                                    "type": "string",
                                                    "example": "Carlos Garcia"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/docentes/{id}": {
            "put": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Actualiza un registro de docente.",
                "operationId": "fca2f373c6b49bd51ba5cca51fb61c0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del docente a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Garcia"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos.garcia@example.com"
                                    },
                                    "identification_number": {
                                        "type": "string",
                                        "example": "987-654321-0"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1980-03-15"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "111-222-3333"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Avenida Siempre Viva 742"
                                    },
                                    "profile_image": {
                                        "description": "Optional: Teacher's profile image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "department": {
                                        "type": "string",
                                        "example": "Matemáticas"
                                    },
                                    "hire_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2010-08-01"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "50000.00"
                                    },
                                    "qualification": {
                                        "type": "string",
                                        "example": "PhD en Matemáticas"
                                    },
                                    "specialization": {
                                        "type": "string",
                                        "example": "Álgebra Lineal"
                                    },
                                    "office_location": {
                                        "type": "string",
                                        "example": "Edificio A, Oficina 201"
                                    },
                                    "emergency_contact_name": {
                                        "type": "string",
                                        "example": "Ana Garcia"
                                    },
                                    "emergency_contact_phone": {
                                        "type": "string",
                                        "example": "555-987-6543"
                                    },
                                    "emergency_contact_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ana.garcia@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Docente actualizado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Teacher updated successfully"
                                        },
                                        "teacher": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre_completo": {
                                                    "type": "string",
                                                    "example": "Carlos Garcia"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Docente no encontrado."
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}